Organize components by feature and update imports
