+++ src/components/features/AuthorBio.astro
@@ -3,8 +3,8 @@
  * AuthorBio component
  * Displays information about a post author
  */
-import ResponsiveImage from './ResponsiveImage.astro';
-import { BLOG_CONFIG } from '../config/current-config';
+import ResponsiveImage from '../ui/ResponsiveImage.astro';
+import { BLOG_CONFIG } from '../../config/current-config';
 
 interface Props {
   name: string;
