+++ src/components/features/RelatedPosts.astro
@@ -4,8 +4,8 @@
  * Displays a list of posts related to the current post
  */
 
-import type { BlogPost } from '../types/content';
-import PostCard from './PostCard.astro';
+import type { BlogPost } from '../../types/content';
+import PostCard from '../ui/PostCard.astro';
 // FormattedDate and ResponsiveImage are now used within PostCard
 
 interface Props {
