+++ src/components/layout/ConfigurableFooter.astro
@@ -2,7 +2,7 @@
 import { BLOG_CONFIG } from '../../config/current-config';
 import SocialLinks from '../ui/SocialLinks.astro';
 import Logo from '../ui/Logo.astro';
-import NewsletterSignup from '../NewsletterSignup.astro';
+import NewsletterSignup from '../features/NewsletterSignup.astro';
 
 const { site, branding, navigation, content, features, social } = BLOG_CONFIG;
 const currentYear = new Date().getFullYear();
