+++ src/components/ui/PostCard.astro
@@ -1,8 +1,8 @@
 --- 
 // src/components/PostCard.astro
-import type { BlogPost } from '../types/content';
+import type { BlogPost } from '../../types/content';
 import ResponsiveImage from './ResponsiveImage.astro';
-import { BLOG_CONFIG } from '../config/current-config';
+import { BLOG_CONFIG } from '../../config/current-config';
 
 export interface Props {
   post: BlogPost;
