+++ src/components/ui/ResponsiveImage.astro
@@ -12,7 +12,7 @@
  * Uses optimized images created during the build process
  * Works with the separated content repository structure
  */
-import { resolveContentImagePath } from '../lib/imagePathResolver';
+import { resolveContentImagePath } from '../../lib/imagePathResolver';
 
 interface Props {
   src: string;
