+++ src/layouts/BaseLayout.astro
@@ -5,7 +5,7 @@ import ConfigurableFooter from '../components/layout/ConfigurableFooter.astro';
 import '../styles/global.css';
 import { BLOG_CONFIG } from '../config/current-config';
 import type { BaseLayoutProps } from '../types/content';
-import { generatePageTitle, generateOgImageUrl } from '../lib/seo/utils';
+import { generatePageTitle, generateOgImageUrl } from '../lib/seo';
 
 const {
   title,
