+++ src/lib/content/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Content-related utilities barrel file
+ */
+export * from './queries';
+export * from './search';
+export { getRelatedPostsLegacy } from './related';
