+++ src/pages/404.astro
@@ -2,7 +2,7 @@
 import { getCollection } from 'astro:content';
 import { compareDesc } from 'date-fns';
 import BaseLayout from '../layouts/BaseLayout.astro';
-import FormattedDate from '../components/FormattedDate.astro';
+import FormattedDate from '../components/ui/FormattedDate.astro';
 
 // Get recent posts for suggestions (safely handling empty collections)
 let recentPosts = [];
