+++ src/pages/categories/[category]/[page].astro
@@ -1,10 +1,10 @@
 ---
 import { getCollection, type CollectionEntry } from 'astro:content';
 import BaseLayout from '../../../layouts/BaseLayout.astro';
-import FormattedDate from '../../../components/FormattedDate.astro';
-import PageNavigation from '../../../components/PageNavigation.astro';
-import ResponsiveImage from '../../../components/ResponsiveImage.astro';
-import PostCard from '../../../components/PostCard.astro';
+import FormattedDate from '../../../components/ui/FormattedDate.astro';
+import PageNavigation from '../../../components/features/PageNavigation.astro';
+import ResponsiveImage from '../../../components/ui/ResponsiveImage.astro';
+import PostCard from '../../../components/ui/PostCard.astro';
 import { SEO } from 'astro-seo';
 import { BLOG_CONFIG } from '../../../config/current-config';
