+++ src/pages/posts/page/[page].astro
@@ -2,8 +2,8 @@
 import { getCollection, type CollectionEntry } from 'astro:content';
 import type { Page } from 'astro';
 import BaseLayout from '../../../layouts/BaseLayout.astro';
-import PageNavigation from '../../../components/PageNavigation.astro';
-import PostCard from '../../../components/PostCard.astro';
+import PageNavigation from '../../../components/features/PageNavigation.astro';
+import PostCard from '../../../components/ui/PostCard.astro';
 // FormattedDate and ResponsiveImage are now used within PostCard
 import { SEO } from 'astro-seo';
 import { BLOG_CONFIG } from '../../../config/current-config';
