+++ src/pages/tags/[tag].astro
@@ -1,6 +1,6 @@
 ---
-import { getAllPosts } from '../../lib/content/queries';
-import PostCard from '../../components/PostCard.astro';
+import { getAllPosts } from '../../lib/content';
+import PostCard from '../../components/ui/PostCard.astro';
 import BaseLayout from '../../layouts/BaseLayout.astro';
 
 export async function getStaticPaths() {
