+++ src/pages/tags/[tag]/[page].astro
@@ -1,9 +1,9 @@
 ---
 import { getCollection, type CollectionEntry } from 'astro:content';
 import BaseLayout from '../../../layouts/BaseLayout.astro';
-import FormattedDate from '../../../components/FormattedDate.astro';
-import PageNavigation from '../../../components/PageNavigation.astro';
-import ResponsiveImage from '../../../components/ResponsiveImage.astro';
+import FormattedDate from '../../../components/ui/FormattedDate.astro';
+import PageNavigation from '../../../components/features/PageNavigation.astro';
+import ResponsiveImage from '../../../components/ui/ResponsiveImage.astro';
 import { SEO } from 'astro-seo';
 import { BLOG_CONFIG } from '../../../config/current-config';
